Follow convention of aligning icon with window
Conventionally on OS X tray windows are lined up with the left of the
icon if the width won't exceed the screen width otherwise they're right
aligned.

This changes the tray manager to align the window with the left of icon
when possible otherwise aligns the right of the window with the right of
the icon.