Add coin settings page
This adds a "bar" at the bottom of the coin tracker that displays your
total worth + a settings button that opens the settings page.

In the settings page you can set the total of each coin that you own and
enable coins to show on/off in the menu bar.